requirements.txt
pytest==6.2.1
requests==2.24.0
behave==1.2.6
selenium==3.141.0
numpy==1.19.3
opencv_python==4.4.0.46
locustio==0.999
PyHamcrest==2.0.2
